INS-1508: add base for slot element
